Add ability to provide a function for the `allowOutsideClick` prop
The focus-trap package allows either a boolean or a callback function to
be provided to its allowOutsideClick, whereas `focus-trap-vue` currently
only allows a boolean to be provided.

Adding the ability to provide a callback function gives developers
access to the `MouseEvent` emitted when a DOM element outside the focus
trap is clicked, allowing them to dynamically determine whether or not
the outside click should be allowed.