fix(props): Upgrade `focus-trap` to fix `initialFocus` prop types
The `focus-trap` library was updated in version 6.7.0 to fix an
issue with the type used for the `initialFocus` option where passing
`false` or a function that returns `false` was not allowed.

The relevant release is tagged here:
https://github.com/focus-trap/focus-trap/releases/tag/v6.7.0

Updating the type to include the Boolean constructor will allow
users of the Vue component to pass the `false` literal for the
`:initial-focus` prop. A function returning `false` already passed
the Vue type check but in previous versions of `focus-trap` would
raise an error. This was also fixed in 6.7.0:

https://github.com/focus-trap/focus-trap/commit/14b0ee830bd6d25fd86bac38135a590f98f43d25#diff-a3ce29c1c993dbf3f968461bb3ff5e3f522d8b0c94cc2ca8f6b3ef7a9eda3621R209-R212

Updating the peer and dev dependencies to be at least 6.7.0
allows `vue-focus-trap` to have the correct behavior and types
and to match the latest documentation for the `focus-trap` library.